Consolidated interaction event scripts
Now all interaction-related event handling code is in one place, along with the tests.
